+++ Source/Shared/Algorithm.swift
@@ -1,49 +1,6 @@
 import Foundation
 
 class Algorithm {
-  enum Counter {
-    case zero, one, many
-
-    func increment() -> Counter {
-      switch self {
-      case .zero:
-        return .one
-      case .one:
-        return .many
-      case .many:
-        return self
-      }
-    }
-  }
-
-  class TableEntry: Equatable {
-    var oldCounter: Counter = .zero
-    var newCounter: Counter = .zero
-    var indexesInOld: [Int] = []
-
-    static func == (lhs: TableEntry, rhs: TableEntry) -> Bool {
-      return lhs.oldCounter == rhs.oldCounter &&
-        lhs.newCounter == rhs.newCounter &&
-        lhs.indexesInOld == rhs.indexesInOld
-    }
-  }
-
-  enum ArrayEntry: Equatable {
-    case tableEntry(TableEntry)
-    case indexInOther(Int)
-
-    public static func == (lhs: ArrayEntry, rhs: ArrayEntry) -> Bool {
-      switch (lhs, rhs) {
-      case (.tableEntry(let l), .tableEntry(let r)):
-        return l == r
-      case (.indexInOther(let l), .indexInOther(let r)):
-        return l == r
-      default:
-        return false
-      }
-    }
-  }
-
   public init() {}
 
   public func diff<T: Hashable>(old: [T], new: [T]) -> [Change<T>] {
@@ -69,12 +26,12 @@ class Algorithm {
     newArray.enumerated().forEach { indexOfNew, item in
       switch item {
       case .tableEntry(let entry):
-        guard !entry.indexesInOld.isEmpty else {
-          return
-        }
+        guard !entry.indexesInOld.isEmpty else { return }
         let indexOfOld = entry.indexesInOld.removeFirst()
-        let isObservation1 = entry.newCounter == .one && entry.oldCounter == .one
-        let isObservation2 = entry.newCounter != .zero && entry.oldCounter != .zero && newArray[indexOfNew] == oldArray[indexOfOld]
+        let isObservation1 = entry.newCounter == .one &&
+          entry.oldCounter == .one
+        let isObservation2 = entry.newCounter != .zero &&
+          entry.oldCounter != .zero && newArray[indexOfNew] == oldArray[indexOfOld]
         if isObservation1 || isObservation2 {
           newArray[indexOfNew] = .indexInOther(indexOfOld)
           oldArray[indexOfOld] = .indexInOther(indexOfNew)
