+++ Source/Shared/TableEntry.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+class TableEntry: Equatable {
+  var oldCounter: Counter = .zero
+  var newCounter: Counter = .zero
+  var indexesInOld: [Int] = []
+
+  static func == (lhs: TableEntry, rhs: TableEntry) -> Bool {
+    return lhs.oldCounter == rhs.oldCounter &&
+      lhs.newCounter == rhs.newCounter &&
+      lhs.indexesInOld == rhs.indexesInOld
+  }
+}
